Move assets folder into app folder and rewrite imports
