+++ src/app/assets/HeartIcon.module.css
