+++ src/app/assets/fonts/Raleway/static/Raleway-ExtraBold.ttf
