+++ src/app/assets/fonts/Raleway/static/Raleway-ExtraBoldItalic.ttf
