+++ src/app/assets/fonts/Raleway/static/Raleway-Italic.ttf
