+++ src/app/components/BackButton/BackButton.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import BackIcon from '../../../assets/BackIcon';
+import BackIcon from '../../assets/BackIcon';
 import styles from './BackButton.module.css';
 
 export default function BackButton(): JSX.Element {
