+++ src/app/components/DogCard/DogCard.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import HeartButton from '../HeartButton/HeartButton';
-import Images from '../Images/Images';
+import Image from '../Image/Image';
 import styles from './DogCard.module.css';
 
 type DogCardProps = {
@@ -14,7 +14,7 @@ export default function DogCard({ image, isLiked }: DogCardProps): JSX.Element {
       <div className={styles.heartButton}>
         <HeartButton onButtonClick={() => console.log('')} isActive={isLiked} />
       </div>
-      <Images image={`${image}`} className={styles.image} />
+      <Image image={`${image}`} className={styles.image} />
     </section>
   );
 }
