+++ src/app/components/SlideShow/SlideShow.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './SlideShow.module.css';
-import Images from '../Images/Images';
-import type { ImagesProps } from '../Images/Images';
+import Images from '../Image/Image';
+import type { ImagesProps } from '../Image/Image';
 
 export default function SlideShow({ image }: ImagesProps): JSX.Element {
   return (
